[HW13] Running server to separate commang
